postgres/requirements.txt
psycopg2-binary==2.8.5
pandas==1.0.4
